moved disble comments functionalities to standalone class
